Setup component for full widget basic functionality dev
