+++ client/components/FavoritePop.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+const FavoritePop = () => (
+
+<div>FavoritePop</div>
+
+
+);
+
+export default RecHome;
